requirements.txt
python-dotenv
flask
anthropic
requests
streamlit
beautifulsoup4
Langchain
serpapi